move count_na_features to cyto_utils
