+++ pycytominer/cyto_utils/features.py
@@ -92,6 +92,23 @@ def infer_cp_features(population_df, metadata=False):
     return features
 
 
+def count_na_features(population_df, features):
+    """
+    Given a population dataframe and features, count how many nas per feature
+
+    Arguments:
+    population_df - pandas DataFrame storing profiles
+    features - a list of features present in the population dataframe
+
+    Return:
+    Dataframe of NA counts per variable
+    """
+
+    return pd.DataFrame(
+        population_df.loc[:, features].isna().sum(), columns=["num_na"]
+    )
+
+
 def drop_outlier_features(
     population_df, features="infer", samples="all", outlier_cutoff=15
 ):
